azure-functions/requirements.txt
azure-functions
torch==2.0.1
transformers==4.30.0
